+++ src/components/Stages/Quizes/QuizTwo.tsx
@@ -1,11 +1,12 @@
-import React, { useCallback, useContext, useState } from "react";
+import { useCallback, useContext, useState } from "react";
 import { GameSettingsContext } from "../../../providers/GameSettingsContext";
 import { RoutingContext } from "../../../providers/RoutingContext";
-import { Quiz } from "../../../stories/forms/QuizForm/QuizForm";
-import { useLockNoPrompt } from "../../../utils/lockNavigation";
 import { PrimaryButton, SkipButton } from "../../../stories/buttons";
+import { Quiz } from "../../../stories/forms/QuizForm/QuizForm";
 import { questionSetTwo } from "../../../QuizSets";
+
 import { DefaultMessages } from "../../../Messages";
+import { useLockNoPrompt } from "../../../utils/lockNavigation";
 
 export const QuizTwo = () => {
   const [unLockNavigation, setUnlockNavigation] = useState<boolean>(true);
