+++ src/providers/GameSettingsContext.tsx
@@ -1,4 +1,4 @@
-import React, { createContext } from "react";
+import { createContext } from "react";
 import { useLocalStorage } from "../utils/useLocalStorage";
 
 const defaultValues = {
