+++ src/providers/RoutingContext.tsx
@@ -1,8 +1,8 @@
-import React, { createContext, useContext } from "react";
+import { createContext, useContext } from "react";
 import { useNavigate } from "react-router-dom";
+import { GameSettingsContext } from "./GameSettingsContext";
 
 import { PrivateRoutes } from "../PrivateRoutes";
-import { GameSettingsContext } from "./GameSettingsContext";
 
 const defaultValues = {
   routeStart: () => {},
