+++ src/stories/buttons/NavigationButton/NavigationButton.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./NavigationButton.css";
 
 interface NavigationButtonProps {
