+++ src/stories/buttons/SubmitButton/SubmitButton.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./SubmitButton.css";
 
 interface SubmitButtonProps {
