+++ src/stories/buttons/UnlockButton/UnlockButton.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./UnlockButton.css";
 
 interface UnlockButtonProps {
