+++ src/stories/timer/CountdownTimer/CountdownTimer.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, FC } from "react";
+import { useState, useEffect, FC } from "react";
 import Countdown, { CountdownRenderProps } from "react-countdown";
 import "./CountdownTimer.css";
